Allow configurable speaker and add cow
The speaker now is not a feature anymore but a configurable
part of the library. The cow from cowsay was added as a speaker
as well.

The speaker is configured using the set_speaker function and the
default speaker is Ferris.

In the binary, the speaker is configured using the -t flag.